Add `handle_error` function to elasticsearch_dynamic
This is a hacky solution to copy part of the behaviour that is lost in
the `elastcsearch_dynamic` output.
It has a subset of the `handle_error` from ElasticsearchErrorHandler,
because the original one wasn't prepared for the ElasticsearchDynamic
payload.
It will now properly emit an error event for each rejection (response
code 400) from ElasticSearch.